Trying to format get requests for Users
In order to login or create a user, you need to check the database to
see if the user exists (for creating an account this is done to prevent
duplicate usernames). A week an a half ago I was able to send a POST
request and create a user; however, I am having trouble formating the
GET request. I tried what you see here; however, Postman suggests that I
format the GET request as http://localhost:5050/users?Username={Username}.
Either way neither of them work.