Fixed login for now
Users are now able to "log in", meaning really the connections work now.
Earlier in the project, where had a very simple js file that stored
whether the user was logged in. We will no longer be going with this and
will likely move to Redux to manage tokens.

I will need to create a Redux store and also need to start managing
JWT.