Redux does not work actually
Redux used to work; however, now it does not. I believe the difference
is that for my test, the counter was synchronous, while using apis and
login details is async, so it seems that we will likely have to use
Redux Thunk as a middleware for async functionality. When I click login,
I do not get the console alert until I refresh the page, and then it is
always "Failed to login user", which sources online seem to say is
because of the aforementioned async problems.

Guides online seem to have the api request to the server handled in the
authSlice file, however, this will require a decent amount of
refactoring. This could also be a complete waste of time for some
unforseeable reason.

The test.jsx is honestly unnecessary, I just like keeping it there as
reference.